build single Edge element

build single Edge element

Usage

buildEdge(
  source = NULL,
  target,
  width = 3,
  curveStyle = "haystack",
  label = "",
  fontSize = 16,
  lineColor = "#FECA57",
  lineStyle = "solid",
  sourceArrowColor = "#feca57",
  targetArrowColor = "#feca57",
  sourceArrowShape = "none",
  targetArrowShape = "none",
  opacity = 1,
  tooltip = ""
)

Arguments

source

edge linked node's id. [string]

target

edge linked target node's id. [string]

width

The width of an edge’s line. [numeric]

curveStyle

The curving method used to separate two or more edges between two nodes. [string]

label

edge's label [string]

fontSize

edge labels font size [numeric]

lineColor

The colour of the edge’s line. [string]

lineStyle

The style of the edge’s line; may be solid, dotted, or dashed. [string]

sourceArrowColor

The colour of the edge’s source arrow. [string]

targetArrowColor

The colour of the edge’s target arrow. [string]

sourceArrowShape

The shape of the edge’s source arrow. [string]

targetArrowShape

The shape of the edge’s target arrow. [string]

opacity

Opacity of edge itself. [numeric between 0 ~ 1]

tooltip

Text for tooltip. [string]

Value

List typed Edge element, consisted with data options ( source, target, data ) and style options ( width, curvestyle... )

See also