buildROptions

build Rendering Option

Usage

buildROptions(
  headless = FALSE,
  styleEnabled = TRUE,
  hideEdgesOnViewport = FALSE,
  textureOnViewport = FALSE,
  motionBlur = FALSE,
  motionBlurOpacity = 0.2,
  wheelSensitivity = 1,
  pixelRatio = "auto"
)

Arguments

headless

A convenience option that initialises the instance to run headlessly. [logical]

styleEnabled

Whether style available or not. [logical]

hideEdgesOnViewport

Whether edge will show on canvas manipulation. [logical]

textureOnViewport

Whether texture used in canvas manipulation. [logical]

motionBlur

Whether use motionBlur effect. [logical]

motionBlurOpacity

opacity of motion blur frames [numeric between 0 ~ 1 (transparent)]

wheelSensitivity

Changes the scroll wheel sensitivity when zooming. [numeric between 0 (zoom slower) ~ 1 (zoom faster)]

pixelRatio

Overrides the screen pixel ratio with a manually set value [numeric]

Value

List typed Rendering Option for Cytoscape.js canvas object.

Details

undescribed parameter will set as default.

See also

Examples

ropt <- buildROptions(wheelSensitivity = 0.5)