cytoscape.js in shiny application

generate canvas with given network element and options

Usage

shinyCyJS(
  elements = list(),
  options = list(),
  layout = list(name = "cose"),
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)

Arguments

elements

node and edge objects, it should be list of element.

options

rendering / interaction options, can be created with buildIoption(), buildRoption()

layout

list type layout, it must be contain name and other optional values

width

canvas width.

height

canvas height.

elementId

id used to identify in application.

...

other parameters