build single node element.

build single node element.

buildNode(
  id = NULL,
  width = 15,
  height = 15,
  shape = "ellipse",
  bgColor = "#48DBFB",
  bgOpacity = 1,
  bgFill = "solid",
  bgBlacken = 0,
  borderWidth = 0,
  borderStyle = "solid",
  borderColor = "#8395a7",
  borderOpacity = 1,
  isParent = FALSE,
  label = NULL,
  labelColor = "#8395a7",
  textOpacity = 1,
  fontSize = 16,
  textOutlineColor = "#222f3e",
  textOutlineOpacity = 1,
  textOutlineWidth = 0,
  textbgColor = "#FFF",
  textbgOpacity = 0,
  textBorderColor = "#222f3e",
  textBorderOpacity = 0,
  textBorderWidth = 0,
  parent = NULL,
  opacity = 1,
  pieSize = rep("0%", 16),
  pieColor = rep("#000", 16),
  tooltip = ""
)

Arguments

id
id of node element. Also it will used as label. [string]

width
Width. [numeric]

height
Height. [numeric]

shape
Shape of node body. polygon not accepted. [string]

bgColor
Background color of node body. [string]

bgOpacity
Opacity of backgroundColor. [numeric between 0 ~ 1]

bgFill
The filling style of the node’s body; may be solid (default), linear-gradient, or radial-gradient. [string]

bgBlacken
Blackens the node’s body for values from 0 to 1; whitens the node’s body for values from 0 to -1. [numeric between -1 ~ 1]

borderWidth
The size of the node’s border. [numeric]

borderStyle
The style of the node’s border; may be solid, dotted, dashed, or double. [string]

borderColor
The colour of the node’s border. [string]

borderOpacity
The opacity of the node’s border. [numeric between 0 ~ 1]

isParent
whether this node is parent node or not [logical]

label
node's label, default is node's id [string]

labelColor
The color of node's label

textOpacity
The opacity of the label text, including its outline. [numeric between 0 ~ 1]

fontSize
The size of the label text. [numeric]

textOutlineColor
The colour of the outline around the element’s label text. [string]

textOutlineOpacity
The opacity of the outline on label text. [numeric between 0 ~ 1]

textOutlineWidth
The size of the outline on label text. [numeric]

textbgColor
colour to apply on the text background. [string]

textbgOpacity
The opacity of the label background; the background is disabled for 0 (default value). [numeric between 0 ~ 1]

textBorderColor
The colour of the border around the label. [string]

textBorderOpacity
The width of the border around the label; the border is disabled for 0 (default value) [numeric between 0 ~ 1]

textBorderWidth
The width of the border around the label. [numeric]

parent
Indicate whith node is parent of this node [string]

opacity
Opacity of node itself. [numeric between 0 ~ 1]

pieSize
Implement for pie node, consisted with 16 pie size[string]

pieColor
Color for each pie part. [string]

tooltip
Text for tooltip. [string]

Value

List typed Node element, consisted with data options ( id ) and style options ( width, shape... )

See also